Renamed Controller classes to DataTableBuilders reflecting what they actually do.
