+++ FocalPointDMSClient/ViewModels/GetEquipmentCommand.cs
@@ -2,7 +2,7 @@
 using System.Windows;
 using System.Windows.Input;
 
-using FocalPointDMSClient.Controllers;
+using FocalPointDMSClient.DataTableBuilders;
 using FocalPointDMSClient.Models.OrmModels;
 
 namespace FocalPointDMSClient.ViewModels
@@ -23,8 +23,8 @@ namespace FocalPointDMSClient.ViewModels
         public void Execute(object parameter)
         {
             var mainViewModel = (MainViewModel) Application.Current.Properties["mainViewModel"];
-            var factory = (ControllerFactory) Application.Current.Properties["controllerFactory"];
-            IDataTableProvider controller = factory.GetInstance(EntityType.Equipment);
+            var factory = (DataTableBuilderFactory) Application.Current.Properties["controllerFactory"];
+            IDataTableBuilder controller = factory.GetInstance(EntityType.Equipment);
             mainViewModel.MainDataTable = controller.BuildTable();
             mainViewModel.StatusTextOutput += mainViewModel.MainDataTable.Rows.Count + " Equipment Items Loaded\n";
